fuzzy matches follow exact matches.
opf_files is populated with the same list twice.  Items in the first list are made visible only if they contain an exact match of the search string.  Items in the second list are made visible only if they do NOT contain an exact match of the search string, but they DO contain a fuzzy match.  This gives the appearance of exact matches sorting to the top.