sorting optimizations
The generic quicksort proved to be too slow for regular use.  Also, the swap function was a bottleneck as it was doing a deep copy of RankedEntrys.  Changed entries to be an array of pointers to RankedEntrys, so swap() now just swaps pointers.  Reduced time on worst case sort of ~3500 items from ~3700ms to ~970ms.